Rename cameraPlugin to FlutterBetterCameraPlugin and rename the namespaces
