+++ android/src/main/java/com/tudutu/plugins/flutterBetterCamera/CameraUtils.java
@@ -1,4 +1,4 @@
-package io.flutter.plugins.camera;
+package com.tudutu.plugins.flutterBetterCamera;
 
 import android.app.Activity;
 import android.content.Context;
@@ -10,7 +10,7 @@ import android.hardware.camera2.CameraMetadata;
 import android.hardware.camera2.params.StreamConfigurationMap;
 import android.media.CamcorderProfile;
 import android.util.Size;
-import io.flutter.plugins.camera.Camera.ResolutionPreset;
+import com.tudutu.plugins.flutterBetterCamera.Camera.ResolutionPreset;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
