+++ android/src/main/java/com/tudutu/plugins/flutterBetterCamera/Constants.java
@@ -1,4 +1,4 @@
-package io.flutter.plugins.camera;
+package com.tudutu.plugins.flutterBetterCamera;
 
 public interface Constants {
