+++ android/src/test/java/io/flutter/plugins/camera/DartMessengerTest.java
@@ -1,4 +1,4 @@
-package io.flutter.plugins.camera;
+package com.tudutu.plugins.flutterBetterCamera;
 
 import static junit.framework.TestCase.assertNull;
 import static org.junit.Assert.assertEquals;
