+++ example/android/app/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-  package="io.flutter.plugins.cameraexample">
+  package="com.tudutu.plugins.flutterBetterCameraexample">
 
   <application
     android:icon="@mipmap/ic_launcher"
