+++ example/android/app/src/main/java/io/flutter/plugins/cameraexample/MainActivity.java
@@ -1,10 +1,10 @@
-package io.flutter.plugins.cameraexample;
+package com.tudutu.plugins.flutterBetterCameraexample;
 
 import androidx.annotation.NonNull;
 import io.flutter.embedding.android.FlutterActivity;
 import io.flutter.embedding.engine.FlutterEngine;
 import io.flutter.embedding.engine.plugins.shim.ShimPluginRegistry;
-import io.flutter.plugins.camera.CameraPlugin;
+import com.tudutu.plugins.flutterBetterCamera.CameraPlugin;
 import io.flutter.plugins.pathprovider.PathProviderPlugin;
 import io.flutter.plugins.videoplayer.VideoPlayerPlugin;
