+++ pubspec.yaml
@@ -1,4 +1,4 @@
-name: camera
+name: flutter_better_camera
 description: Flutter plugin for controlling the camera on Android and IOS, supports camera feed, capturing images, capturing videos, streaming image buffers and has support for all the essential features the camera APIs allow (flash, AE and more)
 version: 0.6.1
 
@@ -13,15 +13,15 @@ dev_dependencies:
     sdk: flutter
   flutter_driver:
     sdk: flutter
-homepage: https://github.com/Lightsnap/flutter-better-camera/
+homepage: https://github.com/tudutu/flutter-better-camera/
 flutter:
   plugin:
     platforms:
       android:
-        package: io.flutter.plugins.camera
-        pluginClass: CameraPlugin
+        package: com.tudutu.plugins.flutterBetterCamera
+        pluginClass: FlutterBetterCameraPlugin
       ios:
-        pluginClass: CameraPlugin
+        pluginClass: FlutterBetterCameraPlugin
 
 environment:
   sdk: ">=2.1.0 <3.0.0"
